Documentation

Index

Lattices

SymmetryReduceBZ.Lattices.check_reduced

SymmetryReduceBZ.Lattices.genlat_BCC

SymmetryReduceBZ.Lattices.genlat_BCT

SymmetryReduceBZ.Lattices.genlat_CUB

SymmetryReduceBZ.Lattices.genlat_FCC

SymmetryReduceBZ.Lattices.genlat_HEX

SymmetryReduceBZ.Lattices.genlat_HXG

SymmetryReduceBZ.Lattices.genlat_MCL

SymmetryReduceBZ.Lattices.genlat_MCLC

SymmetryReduceBZ.Lattices.genlat_OBL

SymmetryReduceBZ.Lattices.genlat_ORC

SymmetryReduceBZ.Lattices.genlat_ORCC

SymmetryReduceBZ.Lattices.genlat_ORCF

SymmetryReduceBZ.Lattices.genlat_ORCI

SymmetryReduceBZ.Lattices.genlat_REC

SymmetryReduceBZ.Lattices.genlat_RECI

SymmetryReduceBZ.Lattices.genlat_RHL

SymmetryReduceBZ.Lattices.genlat_SQR

SymmetryReduceBZ.Lattices.genlat_TET

SymmetryReduceBZ.Lattices.genlat_TRI

SymmetryReduceBZ.Lattices.get_latparams

SymmetryReduceBZ.Lattices.get_recip_latvecs

SymmetryReduceBZ.Lattices.minkowski_reduce

SymmetryReduceBZ.Lattices.reduce_basis!

Plotting

SymmetryReduceBZ.Plotting.plot_2Dconvexhull

SymmetryReduceBZ.Plotting.plot_3Dconvexhull

SymmetryReduceBZ.Plotting.plot_convexhulls

Symmetry

SymmetryReduceBZ.Symmetry.calc_bz

SymmetryReduceBZ.Symmetry.calc_ibz

SymmetryReduceBZ.Symmetry.calc_pointgroup

SymmetryReduceBZ.Symmetry.calc_spacegroup

SymmetryReduceBZ.Symmetry.inhull

SymmetryReduceBZ.Symmetry.make_primitive

SymmetryReduceBZ.Symmetry.mapto_BZ

SymmetryReduceBZ.Symmetry.mapto_IBZ

SymmetryReduceBZ.Symmetry.mapto_unitcell

Utilites

SymmetryReduceBZ.Utilities.affine_trans

SymmetryReduceBZ.Utilities.contains

SymmetryReduceBZ.Utilities.contains

SymmetryReduceBZ.Utilities.edgelengths

SymmetryReduceBZ.Utilities.get_uniquefacets

SymmetryReduceBZ.Utilities.mapto_xyplane

SymmetryReduceBZ.Utilities.remove_duplicates

SymmetryReduceBZ.Utilities.sample_circle

SymmetryReduceBZ.Utilities.sample_sphere

SymmetryReduceBZ.Utilities.shoelace

SymmetryReduceBZ.Utilities.sortpts_perm

Functions

Lattices

SymmetryReduceBZ.Lattices.check_reduced — Method

check_reduced(basis)

Verify a lattice basis is Minkowski reduced

Arguments

basis::AbstractArray{<:Real,2}: the lattice basis given by the columns of a 2x2 or 3x3 matrix.

Returns

Bool: a boolean that indicates if the lattice basis is reduced.

Examples

using SymmetryReduceBZ
basis = [1 0; 0 1]
SymmetryReduceBZ.Lattices.check_reduced(basis)
# output
true

SymmetryReduceBZ.Lattices.genlat_BCC — Function

genlat_BCC(a)

Generate a body-centered cubic lattice.

Arguments

a::Real: the lattice constant

type::String="primitive": the lattice type: "primitive" or "conventional".

Returns

AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

Examples

using SymmetryReduceBZ
a=1
SymmetryReduceBZ.Lattices.genlat_BCC(a)
# output
3×3 Array{Float64,2}:
 -0.5   0.5   0.5
  0.5  -0.5   0.5
  0.5   0.5  -0.5

SymmetryReduceBZ.Lattices.genlat_BCT — Function

genlat_BCT(a,c)

Generate a body-centered tetragonal lattice.

Arguments

a::Real: a lattice constant

c::Real: a lattice constant

type::String="primitive": the lattice type: "primitive" or "conventional".

Returns

AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

Examples

using SymmetryReduceBZ
a=1
c=1.2;
SymmetryReduceBZ.Lattices.genlat_BCT(a,c)
# output
3×3 Array{Float64,2}:
 -0.5   0.5   0.5
  0.5  -0.5   0.5
  0.6   0.6  -0.6

SymmetryReduceBZ.Lattices.genlat_CUB — Function

genlat_CUB(a)

Generate a simple cubic lattice.

Arguments

a::Real: the lattice constant

type::String="primitive": the lattice type: "primitive" or "conventional".

Returns

AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

Examples

using SymmetryReduceBZ
a=1
SymmetryReduceBZ.Lattices.genlat_CUB(a)
# output
3×3 Array{Int64,2}:
 1  0  0
 0  1  0
 0  0  1

SymmetryReduceBZ.Lattices.genlat_FCC — Function

genlat_FCC(a)

Generate a face-centered cubic lattice.

Arguments

a::Real: the lattice constant

type::String="primitive": the lattice type: "primitive" or "conventional".

Returns

AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

Examples

using SymmetryReduceBZ
a=1
SymmetryReduceBZ.Lattices.genlat_FCC(a)
# output
3×3 Array{Float64,2}:
 0.0  0.5  0.5
 0.5  0.0  0.5
 0.5  0.5  0.0

SymmetryReduceBZ.Lattices.genlat_HEX — Function

genlat_HEX(a,c)

Generate a hexagonal lattice.

Arguments

a::Real: a lattice constant

c::Real: a lattice constant

type::String="primitive": the lattice type: "primitive" or "conventional".

Returns

AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

Examples

using SymmetryReduceBZ
a=1
c=1.2;
SymmetryReduceBZ.Lattices.genlat_HEX(a,c)
# output
3×3 Array{Float64,2}:
  0.5       0.5       0.0
 -0.866025  0.866025  0.0
  0.0       0.0       1.2

SymmetryReduceBZ.Lattices.genlat_HXG — Function

genlat_HXG(a)

Generate a 2D hexagonal lattice.

Arguments

a::Real: the lattice constant

type::String="primitive": the lattice type: "primitive" or "conventional".

Returns

AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

Examples

using SymmetryReduceBZ
a=1
SymmetryReduceBZ.Lattices.genlat_HXG(a)
# output
2×2 Array{Float64,2}:
 1.0  -0.5
 0.0   0.866025

SymmetryReduceBZ.Lattices.genlat_MCL — Function

genlat_MCL(a,b,c,α)

Generate a monoclinic lattice.

Arguments

a::Real: a lattice constant

b::Real: a lattice constant

c::Real: a lattice constant

α::Real: a lattice angle in radians less than π/2

type::String="primitive": the lattice type: "primitive" or "conventional".

Returns

AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

Examples

using SymmetryReduceBZ
a=1
b=1.2
c=1.4
α=π/6;
SymmetryReduceBZ.Lattices.genlat_MCL(a,b,c,α)
# output
3×3 Array{Float64,2}:
 1.0  0.0  0.0
 0.0  1.2  1.21244
 0.0  0.0  0.7

SymmetryReduceBZ.Lattices.genlat_MCLC — Function

genlat_MCLC(a,b,c,α)

Generate a base-centered monoclinic lattice.

Arguments

a::Real: a lattice constant

b::Real: a lattice constant

c::Real: a lattice constant

α::Real: a lattice angle in radians less than π/2

type::String="primitive": the lattice type: "primitive" or "conventional".

Returns

AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

Examples

using SymmetryReduceBZ
a=1
b=1.2
c=1.4
α=π/6;
SymmetryReduceBZ.Lattices.genlat_MCLC(a,b,c,α)
# output
3×3 Array{Float64,2}:
 0.5  -0.5  0.0
 0.6   0.6  1.21244
 0.0   0.0  0.7

SymmetryReduceBZ.Lattices.genlat_OBL — Function

genlat_OBL(a,b,θ)

Generate an oblique lattice.

Arguments

a::Real: the lattice constant

b::Real: the lattice constant

θ::Real: the lattice angle

type::String="primitive": the lattice type: "primitive" or "conventional".

Returns

AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

Examples

using SymmetryReduceBZ
a=1
b=1.2
θ=π/3
SymmetryReduceBZ.Lattices.genlat_OBL(a,b,θ)
# output
2×2 Array{Float64,2}:
 1.0  0.6
 0.0  1.03923

SymmetryReduceBZ.Lattices.genlat_ORC — Function

genlat_ORC(a,b,c)

Generate an orthorhombic lattice.

Arguments

a::Real: a lattice constant

b::Real: a lattice constant

c::Real: a lattice constant

type::String="primitive": the lattice type: "primitive" or "conventional".

Returns

AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

Examples

using SymmetryReduceBZ
a=1
b=1.4;
c=1.2;
SymmetryReduceBZ.Lattices.genlat_ORC(a,b,c)
# output
3×3 Array{Float64,2}:
 1.0  0.0  0.0
 0.0  1.4  0.0
 0.0  0.0  1.2

SymmetryReduceBZ.Lattices.genlat_ORCC — Function

genlat_ORCC(a,b,c)

Generate a base-centered orthorhombic lattice.

Arguments

a::Real: a lattice constant

b::Real: a lattice constant

c::Real: a lattice constant

type::String="primitive": the lattice type: "primitive" or "conventional".

Returns

AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

Examples

using SymmetryReduceBZ
a=1
b=1.2;
c=1.4;
SymmetryReduceBZ.Lattices.genlat_ORCC(a,b,c)
# output
3×3 Array{Float64,2}:
  0.5  0.5  0.0
 -0.6  0.6  0.0
  0.0  0.0  1.4

SymmetryReduceBZ.Lattices.genlat_ORCF — Function

genlat_ORCF(a,b,c)

Generate a face-centered orthorhombic lattice.

Arguments

a::Real: a lattice constant

b::Real: a lattice constant

c::Real: a lattice constant

type::String="primitive": the lattice type: "primitive" or "conventional".

Returns

AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

Examples

using SymmetryReduceBZ
a=1
b=1.4;
c=1.2;
SymmetryReduceBZ.Lattices.genlat_ORCF(a,b,c)
# output
3×3 Array{Float64,2}:
 0.0  0.5  0.5
 0.7  0.0  0.7
 0.6  0.6  0.0

SymmetryReduceBZ.Lattices.genlat_ORCI — Function

genlat_ORCI(a,b,c)

Generate a body-centered orthorhombic lattice.

Arguments

a::Real: a lattice constant

b::Real: a lattice constant

c::Real: a lattice constant

type::String="primitive": the lattice type: "primitive" or "conventional".

Returns

AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

Examples

using SymmetryReduceBZ
a=1
b=1.4;
c=1.2;
SymmetryReduceBZ.Lattices.genlat_ORCI(a,b,c)
# output
3×3 Array{Float64,2}:
 -0.5   0.5   0.5
  0.7  -0.7   0.7
  0.6   0.6  -0.6

SymmetryReduceBZ.Lattices.genlat_REC — Function

genlat_REC(a,b)

Generate a rectangular lattice.

Arguments

a::Real: the lattice constant

type::String="primitive": the lattice type: "primitive" or "conventional".

Returns

AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

Examples

using SymmetryReduceBZ
a=1
b=1.2
SymmetryReduceBZ.Lattices.genlat_REC(a,b)
# output
2×2 Array{Float64,2}:
 1.0  0.0
 0.0  1.2

SymmetryReduceBZ.Lattices.genlat_RECI — Function

genlat_RECI(a,b)

Generate a body-centered rectangular lattice.

Arguments

a::Real: the lattice constant

type::String="primitive": the lattice type: "primitive" or "conventional".

Returns

AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

Examples

using SymmetryReduceBZ
a=1
b=1.2
SymmetryReduceBZ.Lattices.genlat_RECI(a,b)
# output
2×2 Array{Float64,2}:
  0.5  0.5
 -0.6  0.6

SymmetryReduceBZ.Lattices.genlat_RHL — Function

genlat_RHL(a,α)

Generate a rhombohedral lattice.

Arguments

a::Real: a lattice constant

α::Real: a lattice angle in radians

type::String="primitive": the lattice type: "primitive" or "conventional".

Returns

AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

Examples

using SymmetryReduceBZ
a=1
α=π/6;
SymmetryReduceBZ.Lattices.genlat_RHL(a,α)
# output
3×3 Array{Float64,2}:
  0.965926  0.965926  0.896575
 -0.258819  0.258819  0.0
  0.0       0.0       0.442891

SymmetryReduceBZ.Lattices.genlat_SQR — Function

genlat_SQR(a)

Generate a square lattice.

Arguments

a::Real: the lattice constant

type::String="primitive": the lattice type: "primitive" or "conventional".

Returns

AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

Examples

using SymmetryReduceBZ
a=1
SymmetryReduceBZ.Lattices.genlat_SQR(a)
# output
2×2 Array{Int64,2}:
 1  0
 0  1

SymmetryReduceBZ.Lattices.genlat_TET — Function

genlat_TET(a,c)

Generate a simple tetragonal lattice.

Arguments

a::Real: a lattice constant

c::Real: a lattice constant

type::String="primitive": the lattice type: "primitive" or "conventional".

Returns

AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

Examples

using SymmetryReduceBZ
a=1
c=1.2;
SymmetryReduceBZ.Lattices.genlat_TET(a,c)
# output
3×3 Array{Float64,2}:
 1.0  0.0  0.0
 0.0  1.0  0.0
 0.0  0.0  1.2

SymmetryReduceBZ.Lattices.genlat_TRI — Function

genlat_TRI(a,b,c,α,β,γ)

Generate a triclinic lattice.

Arguments

a::Real: a lattice constant

b::Real: a lattice constant

c::Real: a lattice constant

α::Real: a lattice angle in radians

β::Real: a lattice angle in radians

γ::Real: a lattice angle in radians

type::String="primitive": the lattice type: "primitive" or "conventional".

Returns

AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

Examples

using SymmetryReduceBZ
a=1
b=1.2
c=1.4
α=π/6;
β=π/3;
γ=π/4;
SymmetryReduceBZ.Lattices.genlat_TRI(a,b,c,α,β,γ)
# output
3×3 Array{Float64,2}:
 1.0  0.848528  0.7
 0.0  0.848528  1.01464
 0.0  0.0       0.663702

SymmetryReduceBZ.Lattices.get_latparams — Method

get_latparams(latvecs)

Calculate the lattice constants and angles of a lattice basis.

Arguments

latvecs::AbstractArray{<:Real,2}: the lattice basis as columns of an array.

Returns

A list where the first element is a list lattice constants (a,b,c) and second lattice angles in radians (α,β,γ).

Examples

using SymmetryReduceBZ
latvecs = [1 0; 0 1]
SymmetryReduceBZ.Lattices.get_latparams(latvecs)
# output
2-element Array{Array{Float64,1},1}:
 [1.0, 1.0]
 [1.5707963267948966, 1.5707963267948966]

SymmetryReduceBZ.Lattices.get_recip_latvecs — Function

get_recip_latvecs(real_latvecs, convention)

Calculate the reciprocal lattice vectors.

Arguments

real_latvecs::AbstractArray{<:Real,2}: the real-space lattice vectors or primitive translation vectors as columns of a 2x2 or 3x3 array.

convention::String="ordinary": the convention used to go between real and reciprocal space. The two conventions are ordinary (temporal) frequency and angular frequency. The transformation from real to reciprocal space is unitary if the convention is ordinary.

Returns

recip_latvecs::Array{<:Real,2} the reciprocal lattice vectors (reciprocal primitive translation vectors) as columns of a 2x2 or 3x3 array.

Examples

using SymmetryReduceBZ
real_latvecs=[1 0 0; 0 1 0; 0 0 1]
convention="angular"
SymmetryReduceBZ.Lattices.get_recip_latvecs(real_latvecs,convention)
# output
3×3 Array{Float64,2}:
 6.28319  0.0      0.0
 0.0      6.28319  0.0
 0.0      0.0      6.28319

SymmetryReduceBZ.Lattices.minkowski_reduce — Method

    minkowski_reduce(basis)

Minkowski reduce a lattice basis. Follows the logic of Fig. 4 in "Low-Dimensional Lattice Basis Reduction Revisited" by Nguyen, 2009.

Arguments

basis::AbstractArray{<:Real,2}: the lattice basis given by the columns of a 2x2 or 3x3 array.

Returns

`rbasis:: the Minkowski reduced lattice basis as columns of an array.

Examples

using SymmetryReduceBZ
basis = [1 2 0; 0 1 0; 0 0 1]
SymmetryReduceBZ.Lattices.minkowski_reduce(basis)
# output
3×3 Array{Int64,2}:
 0  1  0
 0  0  1
 1  0  0

SymmetryReduceBZ.Lattices.reduce_basis! — Method

reduce_basis!(basis,k)

Reduces the kth lattice vector. This is accomplished by locating the lattice point closest to the projection of the kth lattice vector onto the line or plane given by the other lattice vector(s), subtracting the closest lattice point from the kth lattice vector, and reordering the lattice vectors by increasing Euclidean norms.

Arguments

basis::AbstractArray{<:Real,2}: the lattice basis as columns of an array.

k::Int: Keeps track of which lattice vector needs to be reduced.

Returns

basis::AbstractArray{<:Real,2}: the partially reduced lattice basis as columns of an array.

k::Int: The index of the lattice vector that needs to be reduced next.

Examples

using SymmetryReduceBZ
basis = Array([1 2 0; 0 1 0; 3 2 1]')
k=2
SymmetryReduceBZ.Lattices.reduce_basis!(basis,k)
basis
# output
3×3 Array{Int64,2}:
 0  1  3
 1  2  2
 0  0  1

Plotting

SymmetryReduceBZ.Plotting.plot_2Dconvexhull — Method

plot_2Dconvexhull(convexhull, ax, color)

Plot a 2D convex hull

Arguments

convexhull::Chull{<:Real}: a convex hull object.

ax::PyObject: an axes object from matplotlib.

color::String: the face color of the convex hull.

Returns

ax::PyObject: updated ax that includes a plot of the convex hull.

Examples

using SymmetryReduceBZ
real_latvecs = [1 0; 0 1]
convention="ordinary"
atom_types=[0]
atom_pos = Array([0 0]')
coords = "Cartesian"
ibzformat = "convex hull"
primitive=false
ibz = SymmetryReduceBZ.Symmetry.calc_ibz(real_latvecs,atom_types,atom_pos,coords,
	ibzformat,primitive,convention)
fig,ax=subplots(figsize=figaspect(1)*1.5)
color="deepskyblue"
ax = SymmetryReduceBZ.Plotting.plot_2Dconvexhull(ibz,ax,color);
SymmetryReduceBZ.Plotting.plot_2Dconvexhull(bz,ax,color)

SymmetryReduceBZ.Plotting.plot_3Dconvexhull — Method

plot_3Dconvexhull(convexhull, ax, color)

Plot a 3D convex hull

Arguments

convexhull::Chull{<:Real}: a convex hull object.

ax::PyObject: an axes object from matplotlib.

color::String: the face color of the convex hull.

plotrange=false: the range over which to plot the convex hull.

Returns

ax::PyObject: updated ax that includes a plot of the convex hull.

Examples

using SymmetryReduceBZ
real_latvecs = [1 0; 0 1]
convention="ordinary"
atom_types=[0]
atom_pos = Array([0 0]')
coords = "Cartesian"
bzformat = "convex hull"
primitive=false
bz = SymmetryReduceBZ.Symmetry.calc_bz(real_latvecs,atom_types,atom_pos,coords,
    bzformat,primitive,convention)
fig = figure(figsize=figaspect(1)*1.5)
ax = fig.add_subplot(111, projection="3d")
ax = SymmetryReduceBZ.Plotting.plot_3Dconvexhull(bz,ax,"deepskyblue")

SymmetryReduceBZ.Plotting.plot_convexhulls — Function

plot_convexhulls(real_latvecs,atom_types,atom_pos,coords,primitive,
    convention,rtol,atol)

Plot the Brillouin and Irreducible Brillouin zone in 2D or 3D.

Arguments

real_latvecs::AbstractArray{<:Real,2}: the basis of a real-space lattice as columns of an array.

atom_types:AbstractArray{<:Int,1}: a list of atom types as integers.

atom_pos::AbstractArray{<:Real,2}: the positions of atoms in the crystal structure as columns of an array.

coords::String: indicates the positions of the atoms are in "lattice" or "Cartesian" coordinates.

primitive::Bool=false: make the unit cell primitive before calculating the the IBZ if equal to true.

convention::String="ordinary": the convention used to go between real and reciprocal space. The two conventions are ordinary (temporal) frequency and angular frequency. The transformation from real to reciprocal space is unitary if the convention is ordinary.

rtol::Real=sqrt(eps(float(maximum(real_latvecs)))) a relative tolerance for floating point comparisons.

atol::Real=0.0: an absolute tolerance for floating point comparisons.

Returns

(fig,ax): the figure and axes Python objects.

Examples

using SymmetryReduceBZ
real_latvecs = [1 0; .5 1]
atom_types=[0]
coords = "Cartesian"
atom_pos = Array([0 0]')
convention = "ordinary"
(fig,ax)=plot_convexhulls(real_latvecs,atom_types,atom_pos,coords,convention)

Symmetry

SymmetryReduceBZ.Symmetry.calc_bz — Function

calc_bz(real_latvecs,atom_types,atom_pos,coords,bzformat,makeprim,
	convention,rtol,atol)

Calculate the Brillouin zone for the given real-space lattice basis.

Arguments

real_latvecs::AbstractArray{<:Real,2}: the real-space lattice vectors or

primitive translation vectors as columns of a 2x2 or 3x3 array.

atom_types:AbstractArray{<:Int,1}: a list of atom types as integers.

atom_pos::AbstractArray{<:Real,2}: the positions of atoms in the crystal

structure as columns of an array.

coords::String: indicates the positions of the atoms are in "lattice" or

"Cartesian" coordinates.

bzformat::String: the format of the Brillouin zone. Options include

"convex-hull" and "half-space".

makeprim::Bool=false: make the unit cell primitive before calculating the

the BZ if equal to `true`.

convention::String="ordinary": the convention used to go between real and

reciprocal space. The two conventions are ordinary (temporal) frequency and
angular frequency. The transformation from real to reciprocal space is
unitary if the convention is ordinary.

rtol::Real=sqrt(eps(float(maximum(real_latvecs)))) a relative tolerance for

floating point comparisons.

atol::Real=0.0: an absolute tolerance for floating point comparisons.

Returns

bz: the vertices or half-space representation of the Brillouin zone

depending on the value of `vertsOrHrep`.

Examples

real_latvecs = [1 0; 0 1]
convention="ordinary"
atom_types=[0]
atom_pos = Array([0 0]')
coords = "Cartesian"
bzformat = "convex hull"
makeprim=false
SymmetryReduceBZ.Symmetry.calc_bz(real_latvecs,atom_types,atom_pos,coords,
	bzformat,makeprim,convention)
# output
Convex Hull of 4 points in 2 dimensions
Hull segment vertex indices:
[3, 2, 1, 4]
Points on convex hull in original order:

[0.5 0.5; 0.5 -0.5; -0.5 -0.5; -0.5 0.5]

SymmetryReduceBZ.Symmetry.calc_ibz — Function

calc_ibz(real_latvecs,atom_types,atom_pos,coords,ibzformat,makeprim,
	convention,rtol,atol)

Calculate the irreducible Brillouin zone of a crystal structure in 2D or 3D.

Arguments

real_latvecs::AbstractArray{<:Real,2}: the basis of a real-space lattice as columns of an array.

atom_types:AbstractArray{<:Int,1}: a list of atom types as integers.

atom_pos::AbstractArray{<:Real,2}: the positions of atoms in the crystal structure as columns of an array.

coords::String: indicates the positions of the atoms are in "lattice" or "Cartesian" coordinates.

ibzformat::String: the format of the irreducible Brillouin zone. Options include "convex-hull" and "half-space".

convention::String="ordinary": the convention used to go between real and reciprocal space. The two conventions are ordinary (temporal) frequency and angular frequency. The transformation from real to reciprocal space is unitary if the convention is ordinary.

makeprim::Bool=false: make the unit cell primitive before calculating the

the IBZ if true.

rtol::Real=sqrt(eps(float(maximum(real_latvecs)))) a relative tolerance for floating point comparisons.

atol::Real=0.0: an absolute tolerance for floating point comparisons.

Returns

ibz: the irreducible Brillouin zone as a convex hull or intersection of half-spaces.

Examples

using SymmetryReduceBZ
real_latvecs = [1 0; 0 1]
convention="ordinary"
atom_types=[0]
atom_pos = Array([0 0]')
coords = "Cartesian"
ibzformat = "convex hull"
makeprim=false
SymmetryReduceBZ.Symmetry.calc_ibz(real_latvecs,atom_types,atom_pos,coords,
	ibzformat,makeprim,convention)
# output
Convex Hull of 3 points in 2 dimensions
Hull segment vertex indices:
[1, 2, 3]
Points on convex hull in original order:

[0.0 0.0; 0.5 0.0; 0.5 0.5]

SymmetryReduceBZ.Symmetry.calc_pointgroup — Function

calc_pointgroup(latvecs,rtol,atol)

Calculate the point group of lattice in 2D or 3D.

Arguments

latvecs::AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

rtol::Real=sqrt(eps(float(maximum(real_latvecs)))): a relative tolerance for floating point comparisons. It is used to compare lengths of vectors and the volumes of primitive cells.

atol::Real=0.0: an absolute tolerance for floating point comparisons. It is used to compare lengths of vectors and the volumes of primitive cells.

Returns

pointgroup::Array{Array{Float64,2},1}: the point group of the lattice. The operators operate on points in Cartesian coordinates.

Examples

using SymmetryReduceBZ
basis = [1 0; 0 1]
SymmetryReduceBZ.Symmetry.calc_pointgroup(basis)
# output
8-element Array{Array{Float64,2},1}:
 [0.0 -1.0; -1.0 0.0]
 [0.0 -1.0; 1.0 0.0]
 [-1.0 0.0; 0.0 -1.0]
 [1.0 0.0; 0.0 -1.0]
 [-1.0 0.0; 0.0 1.0]
 [1.0 0.0; 0.0 1.0]
 [0.0 1.0; -1.0 0.0]
 [0.0 1.0; 1.0 0.0]

SymmetryReduceBZ.Symmetry.calc_spacegroup — Function

calc_spacegroup(real_latvecs,atom_types,atom_pos,coords,rtol,atol)

Calculate the space group of a crystal structure.

Arguments

real_latvecs::AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

atom_types::AbstractArray{<:Int,1}: a list of atom types as integers.

atom_pos::AbstractArray{<:Real,2}: the positions of atoms in the crystal structure as columns of an array.

coords::String: indicates the positions of the atoms are in "lattice" or "Cartesian" coordinates.

rtol::Real=sqrt(eps(float(maximum(real_latvecs)))) a relative tolerance for floating point comparisons.

atol::Real=0.0: an absolute tolerance for floating point comparisons.

Returns

spacegroup: the space group of the crystal structure. The first element of spacegroup is a list of fractional translations, and the second element is a list of point operators. The translations are in Cartesian coordinates, and the operators operate on points in Cartesian coordinates.

Examples

using SymmetryReduceBZ
real_latvecs = Array([1 0; 2 1]')
atom_types = [0, 1]
atom_pos = Array([0 0; 0.5 0.5]')
coords = "Cartesian"
SymmetryReduceBZ.Symmetry.calc_spacegroup(real_latvecs,atom_types,atom_pos,
	coords)
# output
(Any[[0.0, 0.0], [0.0, 0.0], [0.0, 0.0], [0.0, 0.0], [0.0, 0.0], [0.0, 0.0],
[0.0, 0.0], [0.0, 0.0]], Any[[0.0 -1.0; -1.0 0.0], [0.0 -1.0; 1.0 0.0],
[-1.0 0.0; 0.0 -1.0], [1.0 0.0; 0.0 -1.0], [-1.0 0.0; 0.0 1.0],
[1.0 0.0; 0.0 1.0], [0.0 1.0; -1.0 0.0], [0.0 1.0; 1.0 0.0]])

SymmetryReduceBZ.Symmetry.inhull — Function

inhull(point, chull, rtol, atol)

Check if a point lies within a convex hull (including the boundaries).

Arguments

point::AbstractArray{<:Real,1}: a point in Cartesian coordinates.

chull::Chull{Float64}: a convex hull in 2D or 3D.

rtol::Real=sqrt(eps(float(maximum(flatten(chull.points))))): a relative

tolerance for floating point comparisons. Needed when a point is on the
boundary of the convex hull.

atol::Real=0.0: an absolute tolerance for floating point comparisons.

Returns

inside::Bool: if true, the point lies within the convex hull.

Examples

real_latvecs = [1 0; 0 2]
atomtypes=[0]
atompos=Array([0 0]')
bzformat="convex hull"
coords="Cartesian"
convention="ordinary"
makeprim=false
bz=calc_bz(real_latvecs, atomtypes,atompos,coords,bzformat,makeprim,convention)

point = [0,0]
inhull(point,bz)
# output
true

SymmetryReduceBZ.Symmetry.make_primitive — Function

make_primitive(real_latvecs,atom_types,atom_pos,coords,rtol,atol)

Make a given unit cell primitive.

This is a Julia translation of the function by the same in https://github.com/msg-byu/symlib.

Arguments

real_latvecs::AbstractArray{<:Real,2}: the basis of the lattice as columns of an array.

atom_types::AbstractArray{<:Int,1}: a list of atom types as integers.

atom_pos::AbstractArray{<:Real,2}: the positions of atoms in the crystal structure as columns of an array.

coords::String: indicates the positions of the atoms are in "lattice" or "Cartesian" coordinates.

rtol::Real=sqrt(eps(float(maximum(real_latvecs)))) a relative tolerance for floating point comparisons.

atol::Real=0.0: an absolute tolerance for floating point comparisons.

Returns

prim_types::AbstractArray{<:Int,1}: a list of atom types as integers in the primitive unit cell.

prim_pos::AbstractArray{<:Real,2}: the positions of the atoms in in the crystal structure as columns of an array.

prim_latvecs::AbstractArray{<:Real,2}: the primitive lattice vectors as columns of an array.

Examples

import SymmetryReduceBZ.Lattices: genlat_CUB
import SymmetryReduceBZ.Symmetry: make_primitive
a = 1.0
real_latvecs = genlat_CUB(a)
atom_types = [0,0]
atom_pos = Array([0 0 0; 0.5 0.5 0.5]')
ibzformat = "convex hull"
coords = "Cartesian"
convention = "ordinary"
make_primitive(real_latvecs, atom_types, atom_pos, coords)
# output
([0], [0.0; 0.0; 0.0], [1.0 0.0 0.5; 0.0 1.0 0.5; 0.0 0.0 0.5])

SymmetryReduceBZ.Symmetry.mapto_BZ — Function

mapto_BZ(kpoint,recip_latvecs,inv_latvecs,coords,rtol,atol)

Map a k-point to a translationally equivalent point within the Brillouin zone.

Arguments

kpoint::AbstractArray{<:Real,1}: a single k-point in lattice or Cartesian coordinates.

recip_latvecs::AbstractArray{<:Real,2}: the reciprocal lattice vectors as columns of an array.

inv_latvecs::AbstractArray{<:Real,2}: the inverse matrix of the reciprocal lattice vectors.

coords::String: the coordinates of the given k-point, either "lattice"

or "Cartesian". The *k*-point returned will be in the same coordinates.

rtol::Real=sqrt(eps(float(maximum(recip_latvecs)))): a relative tolerance

for floating point comparisons. Finite precision errors creep in when `pt`
is transformed to lattice coordinates because the transformation requires
calculating a matrix inverse. The components of the *k*-point in lattice
coordinates are checked to ensure that values close to 1 are equal to 1.

atol::Real=0.0: an absolute tolerance for floating point comparisons.

Returns

bz_point::AbtractArray{<:Real,1}: the symmetrically equivalent k-point

within the Brillouin zone in either lattice or Cartesian coordinates,
depending on the coordinates specified.

Examples

import LinearAlgebra: inv
recip_latvecs = [1 0 0; 0 1 0; 0 0 1]
inv_latvecs = inv(recip_latvecs)
kpoint = [2, 3, 2]
coords = "Cartesian"
mapto_BZ(kpoint, recip_latvecs, inv_latvecs, coords)
# output
3-element Array{Real,1}:
 0.0
 0.0
 0.0

SymmetryReduceBZ.Symmetry.mapto_IBZ — Function

mapto_IBZ(kpoint,recip_latvecs,inv_rlatvecs,ibz,pointgroup,coords,rtol,atol)

Map a point to a symmetrically equivalent point within the IBZ.

Arguments

kpoint::AbstractArray{<:Real,1}: a k-point in 2D or 3D in Cartesian

coordinates.

recip_latvecs::AbstractArray{<:Real,2}: the reciprocal lattice vectors as

columns of a an array.

inv_rlatvecs::AbstractArray{<:Real,2}: the inverse of the square array

`recip_latvecs`.

ibz::Chull{Float64}: the irreducible Brillouin zone as as a convex hull

objects from `QHull`.

pointgroup::Array{Array{Float64,2},1}: a list of point symmetry operators

in matrix form that operate on points from the left.

coords::String: the coordinates the k-point is in. Options are "lattice"

and "Cartesian". The *k*-point within the IBZ is returned in the same
coordinates.

rtol::Real=sqrt(eps(float(maximum(recip_latvecs)))): a relative tolerance

for floating point comparisons. The *k*-point is first mapped the unit cell
and `rtol` is used when comparing components of the *k*-point to 1. It is
also used for comparing floats to zero when checking if the point lies
within `ibz`.

atol::Real=0.0: an absolute tolerance for floating point comparisons. This

is used everywhere `rtol` is used.

Returns

rot_point::AbstractArray{Real,1}: a symmetrically equivalent k-point to

`kpoint` within the irreducible Brillouin zone in the same coordinates as
`coords`.

#Examples

import SymmetryReduceBZ.Lattices: get_recip_latvecs
import SymmetryReduceBZ.Symmetry: calc_pointgroup, calc_ibz, mapto_IBZ

real_latvecs = [1 0; 0 2]
convention="ordinary"
recip_latvecs = get_recip_latvecs(real_latvecs,convention)
inv_rlatvecs = inv(recip_latvecs)
atomtypes=[0]
atompos=Array([0 0]')
coords="Cartesian"
ibzformat="convex hull"
makeprim=false

ibz=calc_ibz(real_latvecs, atomtypes,atompos,coords,ibzformat,makeprim,convention)
pg = calc_pointgroup(real_latvecs)
kpoint = [2,3]
ibz_point = mapto_IBZ(kpoint,recip_latvecs,inv_rlatvecs,ibz,pg,coords)
# output
2-element Array{Real,1}:
 0.0
 0.0

SymmetryReduceBZ.Symmetry.mapto_unitcell — Function

mapto_unitcell(pt,latvecs,inv_latvecs,coords,rtol,atol)

Map a point to the first unit cell.

Arguments

pt::AbstractArray{<:Real,1}: a point in lattice or Cartesian coordinates.

latvecs::AbstractArray{<:Real,2}: the basis vectors of the lattice as columns of an array.

inv_latvecs::AbstractArray{<:Real,2}: the inverse of the matrix of that contains the lattice vectors.

coords::String: indicates whether pt is in "Cartesian" or "lattice" coordinates.

rtol::Real=sqrt(eps(float(maximum(inv_latvecs)))): a relative tolerance for floating point comparisons. Finite precision errors creep up when pt is transformed to lattice coordinates because the transformation requires calculating a matrix inverse. The components of the point in lattice coordinates are checked to ensure that values close to 1 are equal to 1.

atol::Real=0.0: an absolute tolerance for floating point comparisons.

Returns

AbstractArray{<:Real,1}: a translationally equivalent point to pt in the first unit cell in the same coordinates.

Examples

using SymmetryReduceBZ
real_latvecs = [0 1 2; 0 -1 1; 1 0 0]
inv_latvecs=inv(real_latvecs)
pt=[1,2,3.2]
coords = "Cartesian"
SymmetryReduceBZ.Symmetry.mapto_unitcell(pt,real_latvecs,inv_latvecs,coords)
# output
3-element Array{Float64,1}:
 0.0
 0.0
 0.20000000000000018

Utilities

SymmetryReduceBZ.Utilities.affine_trans — Method

affine_trans(pts)

Calculate the affine transformation that maps the points to the xy-plane.

Arguments

pts::AbstractArray{<:Real,2}: an array of Cartesian points as the columns of a 2D array. The points must all lie on a plane in 3D.

Returns

M::AbstractArray{<:Real,2}: the affine transformation matrix that operates on points in homogeneous coordinates from the left.

Examples

using SymmetryReduceBZ
pts = [0.5 0.5 0.5; 0.5 -0.5 0.5; -0.5 0.5 0.5; -0.5 -0.5 0.5]'
SymmetryReduceBZ.Utilities.affine_trans(pts)
# output
4×4 Array{Float64,2}:
  0.0  -1.0   0.0  0.5
 -1.0   0.0   0.0  0.5
  0.0   0.0  -1.0  0.5
  0.0   0.0   0.0  1.0

SymmetryReduceBZ.Utilities.contains — Function

contains(pt,pts,rtol,atol)

Check if an array of points contains a point.

Arguments

pt::AbstractArray{<:Real,1}: a 1D array of reals

pts::AbstractArray{<:Real,2}: a 2D array of reals. Coordinates of points are the columns of the array.

rtol::Real=sqrt(eps(float(maximum(pts)))): a relative tolerance for floating point comparisons

atol::Real=0.0: an absolute tolerance for floating point comparisons.

Returns

Bool: a boolean that indicates the presence or absence of pt in pts.

Examples

import SymmetryReduceBZ.Utilities: contains
pts = Array([1 2; 2 3; 3 4; 4 5]')
pt = [1,2]
contains(pt,pts)
# output
true

SymmetryReduceBZ.Utilities.contains — Function

contains(array,arrays,rtol,atol)

Check if an array of arrays contains an array.

Arguments

array: an array of reals or arbitrary dimension.

arrays: a abstract array of reals of arbitrary dimension.

rtol::Real=sqrt(eps(float(maximum(pts)))): a relative tolerance for floating point comparisons.

atol::Real=0.0: an absolute tolerance for floating point comparisons.

Returns

Bool: a boolean that indicates the presence of absence of array in arrays.

Examples

import SymmetryReduceBZ.Utilities: contains
arrays = [[1 2; 2 3], [2 3; 4 5]]
array = [1 2; 2 3]
contains(array, arrays)
# output
true

SymmetryReduceBZ.Utilities.edgelengths — Function

edgelengths(basis,radius)

Calculate the edge lengths of a parallelepiped circumscribed by a sphere.

Arguments

basis::Array{<:Real,2}: a 2x2 or 3x3 matrix whose columns give the parallelogram or parallelepiped directions, respectively.

radius::Real: the radius of the sphere.

rtol::Real=sqrt(eps(float(radius))): a relative tolerace for floating point comparisons.

atol::Real=0.0: an absolute tolerance for floating point comparisons.

Returns

[la,lb,lc]::Array{Float64,1}: a list of parallelepiped lengths.

Examples

using SymmetryReduceBZ
basis=Array([1. 0. 0.; 0. 1. 0.; 0. 0. 1.])
radius=3.0
SymmetryReduceBZ.Utilities.edgelengths(basis,radius)
# output
3-element Array{Float64,1}:
 3.0
 3.0
 3.0

SymmetryReduceBZ.Utilities.get_uniquefacets — Method

get_uniquefacets(ch)

Calculate the unique facets of a convex hull object.

SymmetryReduceBZ.Utilities.mapto_xyplane — Method

function mapto_xyplane(pts)

Map Cartesian points embedded in 3D on a plane to the xy-plane embedded in 2D.

Arguments

pts::AbstractArray{<:Real,2}: Cartesian points in 3D as columns of a 2D array.

Returns

AbstractArray{<:Real,2}: Cartesian points in 2D as columns of a 2D array.

Examples

using SymmetryReduceBZ
pts = [0.5 -0.5 0.5; 0.5 -0.5 -0.5; 0.5 0.5 -0.5; 0.5 0.5 0.5]'
SymmetryReduceBZ.Utilities.mapto_xyplane(pts)
# output
2×4 Array{Float64,2}:
 0.0  1.0  1.0  0.0
 0.0  0.0  1.0  1.0

SymmetryReduceBZ.Utilities.remove_duplicates — Function

remove_duplicates(points,rtol,atol)

Remove duplicate points from an array.

Arguments

points::AbstractArray{<:Real,2}: the points are columns of a 2D array.

rtol::Real=sqrt(eps(float(maximum(flatten(points))))): a relative tolerance for floating point comparisons.

atol::Real=0.0: an absolume tolerance for floating point comparisons.

Returns

uniquepts::AbstractArray{<:Real,2}: a 2D array of unique points as columns.

Examples

using SymmetryReduceBZ
points=Array([1 2; 2 3; 3 4; 1 2]')
SymmetryReduceBZ.Utilities.remove_duplicates(points)
# output
2×3 Array{Int64,2}:
 1  2  3
 2  3  4

SymmetryReduceBZ.Utilities.sample_circle — Function

sample_circle(basis,radius,offset,rtol,atol)

Sample uniformly within a circle centered about a point.

Arguments

basis::Array{<:Real,2}: a 2x2 matrix whose columns are the grid generating vectors.

radius::Real: the radius of the circle.

offset::Array{<:Real,1}=[0.,0.]: the xy-coordinates of the center of the circle.

rtol::Real=sqrt(eps(float(radius))): a relative tolerace for floating point comparisons.

atol::Real=0.0: an absolute tolerance for floating point comparisons.

Returns

pts::Array{Float64,2} a matrix whose columns are sample points in Cartesian coordinates.

Examples

using SymmetryReduceBZ
basis=Array([1. 0.; 0. 1.]')
radius=1.0
offset=[0.,0.]
SymmetryReduceBZ.Utilities.sample_circle(basis,radius,offset)
# output
2×5 Array{Float64,2}:
  0.0  -1.0  0.0  1.0  0.0
 -1.0   0.0  0.0  0.0  1.0

SymmetryReduceBZ.Utilities.sample_sphere — Function

sample_sphere(basis,radius,offset,rtol,atol)

Sample uniformly within a circle centered about a point.

Arguments

basis::Array{<:Real,2}: a 3x3 matrix whose columns are the grid generating vectors.

radius::Real: the radius of the sphere.

offset::Array{<:Real,1}=[0.,0.]: the xy-coordinates of the center of the circle.

rtol::Real=sqrt(eps(float(radius))): a relative tolerace for floating point comparisons.

atol::Real=0.0: an absolute tolerance for floating point comparisons.

Returns

pts::Array{Float64,2} a matrix whose columns are sample points in Cartesian coordinates.

Examples

using SymmetryReduceBZ
basis=Array([1. 0. 0.; 0. 1. 0.; 0. 0. 1.])
radius=1.0
offset=[0.,0.,0.]
SymmetryReduceBZ.Utilities.sample_sphere(basis,radius,offset)
# output
3×7 Array{Float64,2}:
  0.0   0.0  -1.0  0.0  1.0  0.0  0.0
  0.0  -1.0   0.0  0.0  0.0  1.0  0.0
 -1.0   0.0   0.0  0.0  0.0  0.0  1.0

SymmetryReduceBZ.Utilities.shoelace — Method

shoelace(vertices)

Calculate the area of a polygon with the shoelace algorithm.

Arguments

vertices::AbstractArray{<:Real,2}: the xy-coordinates of the vertices of the polygon as the columns of a 2D array.

Returns

<:Real: the area of the polygon.

SymmetryReduceBZ.Utilities.sortpts_perm — Method

function sortpts_perm(pts)

Calculate the permutation vector that sorts Cartesian points embedded in 3D that lie on a plane (counter)clockwise with respect to the average of all points.

Arguments

pts::AbstractArray{<:Real,2}: Cartesian points embedded in 3D that all lie on a plane. The points are columns of a 2D array.

Returns

perm::AbstractArray{<:Real,1}: the permutation vector that orders the points clockwise or counterclockwise.

Examples

using SymmetryReduceBZ
pts = [0.5 -0.5 0.5; 0.5 -0.5 -0.5; 0.5 0.5 -0.5; 0.5 0.5 0.5]'
perm=SymmetryReduceBZ.Utilities.sortpts_perm(pts)
pts[:,perm]
# output
3×4 Array{Float64,2}:
  0.5  0.5   0.5   0.5
  0.5  0.5  -0.5  -0.5
 -0.5  0.5   0.5  -0.5